backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
pillow
opencv-python-headless
numpy
requests
python-barcode
pyjwt
python-dotenv
